Ajuste no nome dos comparators do WF
