+++ sigawf/src/main/java/br/gov/jfrj/siga/wf/vraptor/WfConfiguracaoController.java
@@ -17,14 +17,13 @@ import br.gov.jfrj.siga.cp.model.DpCargoSelecao;
 import br.gov.jfrj.siga.cp.model.DpFuncaoConfiancaSelecao;
 import br.gov.jfrj.siga.cp.model.DpLotacaoSelecao;
 import br.gov.jfrj.siga.cp.model.DpPessoaSelecao;
-import br.gov.jfrj.siga.cp.model.HistoricoSuporte;
 import br.gov.jfrj.siga.cp.model.enm.ITipoDeConfiguracao;
 import br.gov.jfrj.siga.dp.CpOrgaoUsuario;
 import br.gov.jfrj.siga.vraptor.CpConfiguracaoHelper;
 import br.gov.jfrj.siga.vraptor.SigaObjects;
 import br.gov.jfrj.siga.vraptor.Transacional;
 import br.gov.jfrj.siga.wf.bl.Wf;
-import br.gov.jfrj.siga.wf.bl.WfConfiguracaoCacheComparator;
+import br.gov.jfrj.siga.wf.bl.WfConfiguracaoComparator;
 import br.gov.jfrj.siga.wf.dao.WfDao;
 import br.gov.jfrj.siga.wf.model.WfConfiguracao;
 import br.gov.jfrj.siga.wf.model.WfDefinicaoDeProcedimento;
@@ -85,10 +84,10 @@ public class WfConfiguracaoController extends WfController {
 
 		List<WfConfiguracao> listConfig = Wf.getInstance().getConf().buscarConfiguracoesVigentes(config);
 
-		for (CpConfiguracao cfg : listConfig) 
+		for (CpConfiguracao cfg : listConfig)
 			cfg.atualizarObjeto();
 
-		Collections.sort(listConfig, new WfConfiguracaoCacheComparator());
+		Collections.sort(listConfig, new WfConfiguracaoComparator());
 
 		WfTipoDeConfiguracao tpconf = WfTipoDeConfiguracao.getById(idTpConfiguracao);
 		CpConfiguracaoHelper.incluirAtributosDeListagem(result, tpconf, (List<CpConfiguracao>) (List) listConfig);
